Refactored project structure - ParametersPrimitiveTwo now has its own folder for easier clarity
