+++ Code/Bachelor.Thesis.Benchmarking.WebApi/ParametersPrimitiveTwo/IAddUserSession.cs
@@ -1,7 +1,7 @@
 ﻿using Bachelor.Thesis.Benchmarking.ParametersPrimitiveTwo;
 using Synnotech.DatabaseAbstractions;
 
-namespace Bachelor.Thesis.Benchmarking.WebApi.Database;
+namespace Bachelor.Thesis.Benchmarking.WebApi.ParametersPrimitiveTwo;
 
 public interface IAddUserSession : IAsyncSession
 {
